Split IP and EA functions into separate files
